chore: move header into separate UI component
